pyproject.toml
[tool.poetry]
name = "django-pet-project"
version = "0.1.0"
description = ""
authors = ["Alexander <sasha.14999933@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
drf-spectacular = "0.27.0"
django-jazzmin = "^2.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
